item/buy API 추가

+++ web/src/main/java/com/example/web/jpa/repository/item/ItemRepository.java
@@ -3,5 +3,5 @@ package com.example.web.jpa.repository.item;
 import com.example.web.jpa.entity.item.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemRepository extends JpaRepository<Item, Long> {
+public interface ItemRepository extends JpaRepository<Item, Integer> {
 }
